requirements.txt
Flask==2.0.1
google-cloud-storage==1.42.0
google-cloud-pubsub==2.7.1
Pillow==8.3.2
gunicorn==20.1.0